+++ static/app/views/settings/organizationMembers/inviteBanner.spec.tsx
@@ -36,6 +36,40 @@ describe('inviteBanner', function () {
     (browserHistory.push as jest.Mock).mockReset();
   });
 
+  it('render banners with feature flag', function () {
+    const org = TestStubs.Organization({
+      features: ['integrations-gh-invite'],
+    });
+
+    render(
+      <InviteBanner
+        missingMembers={missingMembers}
+        onSendInvite={() => undefined}
+        organization={org}
+      />
+    );
+
+    expect(screen.getByTestId('invite-banner')).toBeInTheDocument();
+    expect(screen.queryAllByTestId('invite-missing-member')).toHaveLength(5);
+    expect(screen.getByText('See all 5 missing members')).toBeInTheDocument();
+  });
+
+  it('does not render banner if no feature flag', function () {
+    const org = TestStubs.Organization({
+      features: [],
+    });
+
+    render(
+      <InviteBanner
+        missingMembers={missingMembers}
+        onSendInvite={() => undefined}
+        organization={org}
+      />
+    );
+
+    expect(screen.queryByTestId('invite-banner')).not.toBeInTheDocument();
+  });
+
   it('does not render banner if no missing members', async function () {
     const org = TestStubs.Organization({
       features: ['integrations-gh-invite'],
